backend/requirements.txt
# Core Django and API dependencies
Django==5.1.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2023.3

# Environment and configuration
python-dotenv==1.1.0
dj-database-url==2.1.0

# HTTP requests and API
requests==2.32.3
urllib3==2.5.0

# Trading and financial APIs
kiteconnect==5.0.1
yfinance==0.2.33
yahoo-fin==0.8.9.1
nsepy==0.8
nsepython==2.97

# Data processing and analysis
# pandas==2.1.4
numpy==1.26.2
scipy==1.16.0
ta==0.11.0

# Task scheduling and background jobs
celery==5.4.0
django-celery-beat==2.7.0
redis==5.0.1
kombu==5.4.2
billiard==4.2.1
vine==5.1.0
amqp==5.3.1

# Web scraping and parsing (compatible versions)
beautifulsoup4==4.12.2
# lxml==4.9.3
# Removed pyppeteer due to urllib3 conflict
# Use selenium-stealth or alternative if needed
selenium==4.34.2
selenium-wire==5.1.0

# Environment and configuration
python-dotenv==1.1.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2

# Security and cryptography
cryptography==45.0.5
certifi==2025.7.14

# Utilities
click==8.1.7
colorama==0.4.6
six==1.16.0
packaging==23.2
typing_extensions==4.14.1
python-dateutil==2.8.2
pytz==2023.3.post1

# Development and debugging
rich==14.0.0
Pygments==2.19.1

# Web server (for production deployment)
uvicorn==0.34.2
gunicorn==22.0.0

# WebSocket support (compatible version)
websockets==10.4
blinker==1.4
webdriver-manager==4.0.1

# PostgreSQL support (Python 3.13 compatible)
psycopg[binary]==3.2.9
psycopg2-binary==2.9.10